eatgo-admin-api/src/main/java/com/lyg/eatgo/application/CategoryService.java
package com.lyg.eatgo.application;

import com.lyg.eatgo.domain.Category;
import com.lyg.eatgo.domain.CategoryRepository;
import com.lyg.eatgo.domain.Region;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CategoryService {

    @Autowired
    private CategoryRepository categoryRepository;

    @Autowired
    public CategoryService(CategoryRepository categoryRepository) {
        this.categoryRepository = categoryRepository;
    }

    public List<Category> getCategories() {
        List<Category> categories = categoryRepository.findAll();

        return categories;
    }

    public Category addCategory(String name) {
        Category category = Category.builder().name(name).build();

        categoryRepository.save(category);

        return category;
    }
}
